+++ spec/directions/bake_numbered_exercise/main_spec.rb
@@ -5,13 +5,13 @@ require 'spec_helper'
 RSpec.describe Kitchen::Directions::BakeNumberedExercise do
   it 'calls v1' do
     expect_any_instance_of(Kitchen::Directions::BakeNumberedExercise::V1).to receive(:bake)
-      .with(exercise: 'exercise1', number: '1.1', suppress_solution_if: false, note_suppressed_solutions: false, cases: false)
+      .with(exercise: 'exercise1', number: '1.1', suppress_solution_if: false, note_suppressed_solutions: false, cases: false, solution_stays_put: false)
     described_class.v1(exercise: 'exercise1', number: '1.1', suppress_solution_if: false, note_suppressed_solutions: false, cases: false)
   end
 
   it 'calls bake_solution_v1' do
     expect_any_instance_of(Kitchen::Directions::BakeNumberedExercise::V1).to receive(:bake_solution)
-      .with(exercise: 'exercise1', number: '1.1', divider: 'abc')
+      .with(exercise: 'exercise1', number: '1.1', divider: 'abc', solution_stays_put: false)
     described_class.bake_solution_v1(exercise: 'exercise1', number: '1.1', divider: 'abc')
   end
 end